sin wave script now changes all blend shapes
