+++ Assets/Prototyping/ManStaticMesh.mesh
