+++ Assets/Prototyping/ManStaticMesh.mesh.meta
