+++ Packages/manifest.json
@@ -2,10 +2,9 @@
   "dependencies": {
     "com.unity.burst": "1.8.8",
     "com.unity.collections": "2.1.4",
-    "com.unity.shadergraph": "14.0.8",
     "com.unity.ide.rider": "3.0.24",
-    "com.unity.project-auditor": "0.10.0",
     "com.unity.render-pipelines.universal": "14.0.8",
+    "com.unity.shadergraph": "14.0.8",
     "com.unity.test-framework.performance": "3.0.1",
     "com.unity.ugui": "1.0.0",
     "com.unity.modules.ai": "1.0.0",
